style(website-creation-modal): adaptive ready
