+++ src/components/MainPage/index.tsx
@@ -6,19 +6,20 @@ import Footer from '../Footer';
 import Guarantee from '../Guarantee/index';
 import Interaction from '../Interaction/index';
 import Portfolio from '../Portfolio';
-//import ProjectDescriptionModal from "../ProjectDescriptionModal";
-//import Popup from '../Popup';
 import Promo from '../Promo';
 import Recommendations from '../Recommendations';
 import StageCreate from '../StageCreate';
 import TechStack from '../TechStack';
+//import ProjectDescriptionModal from "../ProjectDescriptionModal";
+//import Popup from '../Popup';
+import WebsiteCreationModal from '../WebsiteCreationModal';
 import Choice from '../Сhoice';
 import styles from './style.module.scss';
 
 export default function MainPage() {
 	return (
 		<main className={styles.main}>
-			{/* <Modal /> */}
+			<WebsiteCreationModal />
 			{/* <Navbar /> */}
 			<Promo />
 			<ComponentsContainer>
