--- src/components/Modal/ModalItem/interface.ts
@@ -1,5 +1,0 @@
-import { IModalItem } from '@/constants/ModalWindow/type';
-
-export interface IModalItemProps {
-	item: IModalItem;
-}
