--- src/components/Modal/ModalItem/style.module.scss
@@ -1,57 +1,0 @@
-.modal-item {
-	display: flex;
-	flex-direction: column;
-	margin-bottom: 14px;
-
-	&__wrappper {
-		display: flex;
-		margin-bottom: 14px;
-
-		>span {
-			color: var(--color-red);
-		}
-	}
-
-	&__title {
-		font: var(--font-button-rg);
-		line-height: 120%;
-	}
-
-	&__item {
-		border: none;
-		font: var(--font-description-main);
-		border-radius: 10px;
-		padding: 17px 10px;
-		background-color: var(--color-gray-bg);
-		line-height: var(--line-height-big);
-		letter-spacing: var(--font-description-letter);
-
-		&:focus-visible {
-			outline: none;
-		}
-
-		&_message {
-			resize: none;
-			padding: 13px 10px;
-			
-			&::placeholder {
-				line-height: 61px;
-				vertical-align: bottom;
-			}
-		}
-	}
-
-	&:last-of-type {
-		margin-bottom: 40px;
-	}
-}
-
-.dropdown {
-	&__wrapper {
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-		cursor: pointer;
-		padding: 17px;
-	}
-}
