+++ src/components/WebsiteCreationModal/ModalItem/interface.ts
@@ -1,0 +1,5 @@
+import { IModalItem } from '@/constants/WebsiteCreationModal/type';
+
+export interface IModalItemProps {
+	item: IModalItem;
+}
