+++ src/components/WebsiteCreationModal/style.module.scss
@@ -1,0 +1,133 @@
+.modal {
+	position: relative;
+	display: flex;
+	flex-direction: column;
+	row-gap: 30px;
+	z-index: 300;
+	background-color: var(--color-bg-white);
+	border-radius: 20px;
+	padding: 28px 40px;
+	max-width: 608px;
+	min-height: 787px;
+
+	&__close-btn {
+		position: absolute;
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		top: 17px;
+		right: 20px;
+		width: 40px;
+		height: 40px;
+		border-radius: 100%;
+		background-color: var(--color-lilac);
+		border: none;
+
+		transition: background-color 0.3s ease-in-out;
+
+		&:hover {
+			background-color: #a6aeff;
+		}
+
+		&:active {
+			background-color: #ccd1fd;
+		}
+
+		&:focus-visible {
+			outline: none;
+		}
+	}
+
+	&__title {
+		font: var(--font-block-title);
+		font-size: var(--text-lg);
+	}
+
+	&__form {
+		display: flex;
+		flex-direction: column;
+		row-gap: 16px;
+	}
+
+	&__btn {
+		font: var(--font-button-rg);
+		color: var(--color-text);
+		padding: 15px 25px;
+		background-image: var(--color-gradient-button);
+		background-size: 400%;
+		background-position: 99%;
+		border-radius: 5px;
+		border: none;
+
+		transition: 0.3s background-position ease-in-out;
+
+		&:focus {
+			outline: none;
+		}
+
+		&:hover {
+			background-position: 50%;
+		}
+
+		&:active {
+			transition: 0.1s background-position ease-in-out;
+			background-position: 10%;
+		}
+	}
+}
+
+@media screen and (min-width: 834px) and (max-width: 1279px) {
+	.modal {
+		max-width: 73%;
+		max-height: 942px;
+		padding: 68px 40px;
+		row-gap: 46px;
+
+		&__close-btn {
+			width: 46px;
+			height: 46px;
+		}
+
+		&__title {
+			font-size: var(--text-xl);
+			word-spacing: -1px;
+		}
+
+		&__form {
+			row-gap: 32px;
+		}
+	}
+}
+
+@media screen and (min-width: 694px) and (max-width: 833px) {
+	.modal {
+		&__title {
+			font-size: var(--text-xl);
+		}
+	}
+}
+
+@media screen and (min-width: 480px) and (max-width: 833px) {
+	.modal {
+		&__title {
+			font-size: var(--text-xl);
+		}
+	}
+}
+
+@media screen and (max-width: 833px) {
+	.modal {
+		max-height: 974px;
+		max-width: 82%;
+		padding: 74px 20px 48px;
+
+		&__close-btn {
+			width: 48px;
+			height: 48px;
+		}
+
+		&__form {
+			row-gap: 36px;
+		}
+	}
+}
