+++ src/constants/WebsiteCreationModal/index.ts
@@ -6,6 +6,7 @@ export const data: IModalItem[] = [
 		title: 'Имя',
 		placeholder: 'Иван Иванов',
 		type: 'text',
+		required: false,
 	},
 	{
 		id: '3bda61fa-12de-4fb7-94bf-c905904a6f97',
@@ -17,21 +18,24 @@ export const data: IModalItem[] = [
 	{
 		id: 'c4edfd5e-2a4e-46d8-8006-3c52e2285e0e',
 		title: 'К вашему номеру подключен Viber, WhatsApp, Telegram',
-		placeholder: 'Да',
+		// placeholder: 'Да',
 		type: 'dropdown',
+		options: ['Да'],
 		required: true,
 	},
 	{
 		id: '09e1eaa7-e356-456a-8bf2-1c021e54a6b6',
 		title: 'Выбор сайта',
-		placeholder: 'Лендинг',
+		// placeholder: 'Лендинг',
 		type: 'dropdown',
+		options: ['Лендинг'],
 		required: true,
 	},
 	{
 		id: '27c81fc6-5e1d-4ea3-b6fb-b3047c9918da',
 		title: 'Дополнительная информация',
 		placeholder: 'Сообщение (необязательно)',
-		type: 'textarea',
+		type: 'text',
+		required: false,
 	},
 ];
