+++ src/constants/WebsiteCreationModal/type.ts
@@ -1,7 +1,8 @@
 export interface IModalItem {
 	id: string;
 	title: string;
-	placeholder: string;
+	placeholder?: string;
 	type: string;
 	required?: boolean;
+	options?: string[];
 }
