+++ src/constants/index.ts
@@ -2,8 +2,8 @@ import './Choice/index';
 import './Contacts/index';
 import './Guarantee/index';
 import './Interaction/index';
-import './ModalWindow/index';
 import './Portfolio/index';
 import './Recommendations/index';
 import './Services';
 import './TechStack/index';
+import './WebsiteCreationModal/index';
